refactor(tags): convert TagLinkButton to TSX with an extensible linkItems list
TagLinkButton built its link contents (icon + label) inline in `view`,
giving extensions no clean seam to add to a tag link short of overriding
the whole component or mutating the returned vnode (fragile — the returned
Link's children aren't reliably a mutable array at extend time).

Convert the component to TypeScript and move the link contents into a
`linkItems()` ItemList rendered via `listItems()` inside the Link, the same
pattern core uses elsewhere. Extensions can now add to a tag link with
`extend('ext:flarum/tags/forum/components/TagLinkButton', 'linkItems', ...)`.

No behaviour change: the icon and label render as before, and the children
passed by addTagList (the tag name, read by SelectDropdown to label the
active item) are untouched.
